Created UserController and Auth.js for common functions
