+++ auth.js
@@ -1,0 +1,29 @@
+const jwt = require('jsonwebtoken')
+
+const authenticateToken = (req, res, callback) => {
+  const authHeader = req.headers['authorization']
+  const token = authHeader && authHeader.split(' ')[1]
+
+  if (!token) {
+    return res.sendStatus(401)
+  }
+
+  jwt.verify(token, process.env.ACCESS_TOKEN_SECRET, (err, user) => {
+    if (err) {
+      return res.sendStatus(403)
+    }
+
+    req.user = user
+    callback()
+  })
+}
+
+const generateAccessToken = (user) => {
+  return jwt.sign({ userId: user._id }, process.env.ACCESS_TOKEN_SECRET, { expiresIn: '15m' })
+}
+
+const generateRefreshToken = (user) => {
+  return jwt.sign({ userId: user._id }, process.env.REFRESH_TOKEN_SECRET)
+}
+
+module.exports = { authenticateToken, generateAccessToken, generateRefreshToken }
